requirements.txt
tensorflow
librosa
matplotlib
PyYAML
